Partial solution for toaster overlay
TOAST.LENGTH_SHORT controls the duration of the message to display.

But this duration is 2.5 seconds long and there isn't any built-in constant to reduce the duration.

We can keep a handler to reduce the duration to 1 second which is intended. But there is some issue in the toaster delay. It's working fine for the first few seconds and then it jumps to one.